@Version 	private Integer version;

+++ src/main/java/ua/org/tumakha/spd/entity/User.java
@@ -108,11 +108,11 @@ public class User {
 		this.userId = userId;
 	}
 
-	public Long getVersion() {
+	public Integer getVersion() {
 		return version;
 	}
 
-	public void setVersion(Long version) {
+	public void setVersion(Integer version) {
 		this.version = version;
 	}
